Use NF::new instead of NF::unchecked_new
As @SergiusIW pointed out, you can cause floats to overflow from
reals to Infinity by adding or multiplying sufficiently large
reals together.

Therefore, we can't avoid the cost of a check at the end of the
computation, to ensure that what we have is still within the
bounds of the condition.