Implement From<NoisyFloat> for f32 and f64.
Unfortunately it's not possible under Rust's current coherence rules to
implement From<NoisyFloat<F, C>> for generic F (it gives error E0210).
But we can still provide the implementation for concrete types like f32
and f64 (and most of the time, F will actually be either f32 or f64).
That's what this commit does.

This commit also implements the widening conversion
From<NoisyFloat<f32, C>> for f64, to be consistent with the Rust
standard library, which also implements From<f32> for f64.